refactor(admin): address reviewer feedback and restore gitignore entries
